Extract volumising and screenshoting logic from bake.js
This separates the UI and core logic for these processes, and allows for
the code to be run on new screenshot.html and volume.html pages.
